+++ R/add_bibtex.R
@@ -10,7 +10,7 @@
 #' @import stringr
 #' @importFrom xfun read_utf8
 #' @import rvest
-#' @importFrom xml2 read_html
+#' @import xml2
 #' @export
 
 
@@ -93,12 +93,12 @@ add_kaggle <- function(url = ''){
     alias <- stringr::str_c(first_name,year)
 
     output <- glue::glue("@online{<<alias>>,
-  author = {<<author>>},
-  title = {<<title>>},
-  year = <<year>>,
-  howpublished = {<<howpublished>>},
-  url = {<<url>>},
-  urldate = {<<urldate>>}
+        author = {<<author>>},
+        title = {<<title>>},
+        year = <<year>>,
+        howpublished = {<<howpublished>>},
+        url = {<<url>>},
+        urldate = {<<urldate>>}
   }"
     ,.open = "<<"
     ,.close = ">>"
@@ -120,39 +120,40 @@ add_wechat <- function(url = ''){
   text <- xml2::read_html(url)
 
   author <- text %>%
-    html_nodes('.rich_media_meta_text') %>%
-    html_text() %>%
-    .[!str_detect(.,'：')] %>%
-    str_trim() %>%
-    str_flatten('') %>%
-    str_extract("[\\p{Han}A-z\\s|:：——，]+")
+    rvest::html_nodes('.rich_media_meta_text') %>%
+    rvest::html_text() %>%
+    .[!stringr::str_detect(.,'：')] %>%
+    stringr::str_trim() %>%
+    stringr::str_flatten('') %>%
+    stringr::str_extract("[\\p{Han}A-z\\s|:：——，]+")
 
   title <- text %>%
-    html_nodes('#activity-name') %>%
-    html_text() %>%
-    str_extract("[\\p{Han}A-z\\s|:——]+") %>%
-    str_trim()
+    rvest::html_nodes('#activity-name') %>%
+    rvest::html_text() %>%
+    stringr::str_extract("[\\p{Han}A-z\\s|:——]+") %>%
+    stringr::str_trim()
 
   year <- Sys.Date() %>% str_sub(1,4)
 
   howpublished <- text %>%
-    html_nodes('#js_name') %>%
-    html_text() %>%
-    str_trim()
+    rvest::html_nodes('#js_name') %>%
+    rvest::html_text() %>%
+    stringr::str_trim()
 
   urldate <- Sys.Date()
 
   first_name <- stringr::str_replace_all(author,'[\\s，]','_')
   alias <- stringr::str_c(first_name,year)
 
-  output <- glue::glue("@online{<<alias>>,
-    author = {<<author>>},
-    title = {<<title>>},
-    year = <<year>>,
-    howpublished = {<<howpublished>>},
-    url = {<<url>>},
-    urldate = {<<urldate>>}
-    }"
+  output <- glue::glue(
+"@online{<<alias>>,
+  author = {<<author>>},
+  title = {<<title>>},
+  year = <<year>>,
+  howpublished = {<<howpublished>>},
+  url = {<<url>>},
+  urldate = {<<urldate>>}
+}"
                        ,.open = "<<"
                        ,.close = ">>"
   )
@@ -173,12 +174,12 @@ add_datacamp <- function(url = ''){
   text <- xml2::read_html(url)
 
   author <- text %>%
-    html_nodes('.course__instructor-name') %>%
-    html_text()
+    rvest::html_nodes('.course__instructor-name') %>%
+    rvest::html_text()
 
   title <- text %>%
-    html_nodes('.header-hero__title') %>%
-    html_text()
+    rvest::html_nodes('.header-hero__title') %>%
+    rvest::html_text()
 
   year <- Sys.Date() %>% str_sub(1,4)